git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/fsfs-format7@1476675 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/svn/file-merge.c
@@ -57,7 +57,7 @@ struct file_merge_baton {
   apr_file_t *original_file;
   apr_file_t *modified_file;
   apr_file_t *latest_file;
-  
+
   /* Counters to keep track of the current line in each file. */
   svn_linenum_t current_line_original;
   svn_linenum_t current_line_modified;
@@ -307,7 +307,7 @@ static int
 get_term_width(void)
 {
   char *columns_env;
-#ifdef TIOCGWINSZ	
+#ifdef TIOCGWINSZ
   int fd;
 
   fd = open("/dev/tty", O_RDONLY, 0);
@@ -367,12 +367,12 @@ prepare_line_for_display(const char *line, apr_pool_t *pool)
 {
   svn_stringbuf_t *buf = svn_stringbuf_create(line, pool);
   size_t width;
-  int line_width = LINE_DISPLAY_WIDTH;
+  size_t line_width = LINE_DISPLAY_WIDTH;
   apr_pool_t *iterpool;
 
   /* Trim EOL. */
   if (buf->len >= 2 &&
-      buf->data[buf->len - 2] == '\r' && 
+      buf->data[buf->len - 2] == '\r' &&
       buf->data[buf->len - 1] == '\n')
     svn_stringbuf_chop(buf, 2);
   else if (buf->len >= 1 &&
@@ -591,7 +591,7 @@ merge_chunks(apr_array_header_t **merged_chunk,
                                                   : chunk2->nelts;
   *abort_merge = FALSE;
 
-  /* 
+  /*
    * Prepare the selection prompt.
    */
 
@@ -645,7 +645,7 @@ merge_chunks(apr_array_header_t **merged_chunk,
         }
       else
         line2 = prepare_line_for_display("", iterpool);
-        
+
       prompt_line = apr_psprintf(iterpool, "%s|%s\n", line1, line2);
 
       svn_stringbuf_appendcstr(prompt, prompt_line);
@@ -910,7 +910,7 @@ svn_cl__merge_file(const char *base_path,
                 svn_dirent_local_style(svn_dirent_skip_ancestor(path_prefix,
                                                                 wc_path),
                                        scratch_pool)));
-                
+
       return SVN_NO_ERROR;
     }
 
@@ -950,6 +950,6 @@ svn_cl__merge_file(const char *base_path,
               svn_dirent_local_style(svn_dirent_skip_ancestor(path_prefix,
                                                               wc_path),
                                      scratch_pool)));
-                
+
   return SVN_NO_ERROR;
 }
